+++ src/icqface.cc
@@ -401,7 +401,7 @@ int icqface::contextmenu(icqcontact *c) {
 
     m.scale();
     m.idle = &menuidle;
-    i = (int) m.getref(m.open()-1);
+    i = (long) m.getref(m.open()-1);
     m.close();
 
     if(i) lastr = i;
@@ -450,7 +450,7 @@ int icqface::generalmenu() {
 
     if(i) {
 	lastitem = i-1;
-	r = (int) m.getref(lastitem);
+	r = (long) m.getref(lastitem);
     }
 
     return r;
@@ -513,7 +513,7 @@ icqcontact *icqface::mainloop(int &action) {
 	    g = 0;
 	}
 	
-	if((unsigned int) c < 100) c = 0;
+	if((unsigned long) c < 100) c = 0;
 
 	if(i) {
 	    switch(action = extk) {
@@ -1505,7 +1505,7 @@ bool icqface::changestatus(vector<protocolname> &pnames, imstatus &st) {
 	m.scale();
 
 	i = m.open();
-	choice = (int) m.getref(i-1);
+	choice = (long) m.getref(i-1);
 
 	m.close();
     }
@@ -1560,7 +1560,7 @@ bool icqface::changestatus(vector<protocolname> &pnames, imstatus &st) {
 	m.close();
 
 	if(r = i) {
-	    st = (imstatus) ((int) m.getref(i-1));
+	    st = (imstatus) ((long) m.getref(i-1));
 	}
     }
 
@@ -2153,7 +2153,7 @@ void icqface::quickfind(verticalmenu *multi) {
 			    c = (icqcontact *) cm->getref(i);
 			}
 
-			if((unsigned int) c > 100) {
+			if((unsigned long) c > 100) {
 			    string current = c->getdispnick();
 			    len = current.size();
 			    if(len > nick.size()) len = nick.size();
